TextField extension was moved. StackView leading and trailing constraints
