+++ ISICbalance/Extensions/TextFieldExt.swift
@@ -1,0 +1,18 @@
+//
+//  TextFieldExt.swift
+//  ISICbalance
+//
+//  Created by Rostislav Babáček on 16/03/2019.
+//  Copyright © 2019 Rostislav Babáček. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension UITextField {
+    func setLeftPaddingPoints(_ amount: CGFloat) {
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: amount, height: self.frame.size.height))
+        self.leftView = paddingView
+        self.leftViewMode = .always
+    }
+}
